+++ app/src/main/java/com/mojilab/moji/ui/main/upload/addCourse/AddCourseActivity.java
@@ -1,14 +1,14 @@
-package com.mojilab.moji.ui.main.upload;
+package com.mojilab.moji.ui.main.upload.addCourse;
 
 import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import com.mojilab.moji.R;
 
-public class TagActivity extends AppCompatActivity {
+public class AddCourseActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_tag);
+        setContentView(R.layout.activity_add_course);
     }
 }
