+++ app/src/main/java/com/mojilab/moji/ui/main/upload/change/ChangeOrderActivity.java
@@ -1,14 +1,14 @@
-package com.mojilab.moji.ui.main.upload;
+package com.mojilab.moji.ui.main.upload.change;
 
 import androidx.appcompat.app.AppCompatActivity;
 import android.os.Bundle;
 import com.mojilab.moji.R;
 
-public class AddCourseActivity extends AppCompatActivity {
+public class ChangeOrderActivity extends AppCompatActivity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_add_course);
+        setContentView(R.layout.activity_change_order);
     }
 }
