+++ app/src/main/res/layout/activity_add.xml
@@ -10,7 +10,7 @@
             android:orientation="vertical"
             android:layout_width="match_parent"
             android:layout_height="match_parent"
-            tools:context=".ui.main.upload.AddActivity">
+            tools:context=".ui.main.upload.add.AddActivity">
 
         <RelativeLayout
                 android:paddingHorizontal="16dp"
